Mention case with owners
Github allows getting all the collaborators that are in the same team as developer who is adding repository. Even though owners have access on Github, they are not part of the team, that is why they are not visible for developer who is trying to add them.
 
In order to solve your problem, you have to add owner to a same team where developer, who adds repository to Semaphore, is.